dw2102: resync fifo when demod locks
If the streaming_ctrl is called to enable TS before demod has locked
the TS will be empty. Antti Palosaari gave me the tip to look at what
was done in the dvbsky driver. Copied the same solution for the
TechnoTrend S2-4600 device: when the state changes from unlock to
lock, call stream_ctrl again.